lib: prepare for compacting of consecutive references
To gain direct access to the nums array, use
`elem.counter.at(elem.location())` instead of using
the `show ref` rule.

This commit relies on the figure.level proposal (typst/typst#6767).

It also serves to demonstrate the subtle issues that arries
with the original behaviour of figure.counter. In particular,
contrast example/example-subpar.typ (orig behaviour) with
example/example-subfig.typ (new behaviour). The reference to
subfigure "1b" is broken in example-subpar.typ but works as
expected in example-subfig.typ.